refactor(map): isolate tiled lifecycle helpers
